+++ src/canvas/StaticCanvas.ts
@@ -6,7 +6,6 @@ import { createCollectionMixin } from '../Collection';
 import { CommonMethods } from '../CommonMethods';
 import type { Pattern } from '../Pattern';
 import { Point } from '../Point';
-import type { BaseFabricObject as FabricObject } from '../EventTypeDefs';
 import type { TCachedFabricObject } from '../shapes/Object/Object';
 import type { Rect } from '../shapes/Rect';
 import type {
@@ -44,6 +43,7 @@ import {
 } from '../util/typeAssertions';
 import { StaticCanvasDOMManager } from './DOMManagers/StaticCanvasDOMManager';
 import type { CSSDimensions } from './DOMManagers/util';
+import type { FabricObject } from '../shapes/Object/FabricObject';
 
 export type TCanvasSizeOptions = {
   backstoreOnly?: boolean;
@@ -336,7 +336,7 @@ export class StaticCanvas<
   }
 
   _onObjectAdded(obj: FabricObject) {
-    if (obj.canvas && obj.canvas !== this) {
+    if (obj.canvas && (obj.canvas as StaticCanvas) !== this) {
       /* _DEV_MODE_START_ */
       console.warn(
         'fabric.Canvas: trying to add an object that belongs to a different canvas.\n' +
@@ -408,7 +408,7 @@ export class StaticCanvas<
    * @param {Boolean}       [options.cssOnly=false]       Set the given dimensions only as css dimensions
    * @deprecated will be removed in 7.0
    */
-  setWidth(value: number, options: TCanvasSizeOptions) {
+  setWidth(value: number, options: TCanvasSizeOptions = {}) {
     return this.setDimensions({ width: value }, options);
   }
 
@@ -420,7 +420,7 @@ export class StaticCanvas<
    * @param {Boolean}       [options.cssOnly=false]       Set the given dimensions only as css dimensions
    * @deprecated will be removed in 7.0
    */
-  setHeight(value: number, options: TCanvasSizeOptions) {
+  setHeight(value: number, options: TCanvasSizeOptions = {}) {
     return this.setDimensions({ height: value }, options);
   }