+++ src/parser/parseAttributes.ts
@@ -1,4 +1,3 @@
-//@ts-nocheck
 import { DEFAULT_SVG_FONT_SIZE } from '../constants';
 import { parseUnit } from '../util/misc/svgParsing';
 import { cPath, fSize, svgValidParentsRegEx } from './constants';
@@ -8,6 +7,7 @@ import { normalizeValue } from './normalizeValue';
 import { parseFontDeclaration } from './parseFontDeclaration';
 import { parseStyleAttribute } from './parseStyleAttribute';
 import { setStrokeFillOpacity } from './setStrokeFillOpacity';
+import type { CSSRules } from './typedefs';
 
 /**
  * Returns an object of attributes' name/value, given element and an array of attribute names;
@@ -17,59 +17,58 @@ import { setStrokeFillOpacity } from './setStrokeFillOpacity';
  * @return {Object} object containing parsed attributes' names/values
  */
 export function parseAttributes(
-  element: SVGElement | HTMLElement,
+  element: HTMLElement | null,
   attributes: string[],
-  svgUid?: string
+  cssRules?: CSSRules
 ): Record<string, any> {
   if (!element) {
     return {};
   }
 
   let value,
-    parentAttributes = {},
+    parentAttributes: Record<string, string> = {},
     fontSize,
-    parentFontSize;
+    parentFontSize = DEFAULT_SVG_FONT_SIZE;
 
-  if (typeof svgUid === 'undefined') {
-    svgUid = element.getAttribute('svgUid');
-  }
   // if there's a parent container (`g` or `a` or `symbol` node), parse its attributes recursively upwards
   if (
     element.parentNode &&
     svgValidParentsRegEx.test(element.parentNode.nodeName)
   ) {
-    parentAttributes = parseAttributes(element.parentNode, attributes, svgUid);
+    parentAttributes = parseAttributes(
+      element.parentElement,
+      attributes,
+      cssRules
+    );
+    if (parentAttributes.fontSize) {
+      fontSize = parentFontSize = parseUnit(parentAttributes.fontSize);
+    }
   }
 
-  let ownAttributes = attributes.reduce(function (memo, attr) {
-    value = element.getAttribute(attr);
-    if (value) {
-      // eslint-disable-line
-      memo[attr] = value;
-    }
-    return memo;
-  }, {});
-  // add values parsed from style, which take precedence over attributes
-  // (see: http://www.w3.org/TR/SVG/styling.html#UsingPresentationAttributes)
-  const cssAttrs = Object.assign(
-    getGlobalStylesForElement(element, svgUid),
-    parseStyleAttribute(element)
-  );
-  ownAttributes = Object.assign(ownAttributes, cssAttrs);
-  if (cssAttrs[cPath]) {
-    element.setAttribute(cPath, cssAttrs[cPath]);
+  const ownAttributes: Record<string, string> = {
+    ...attributes.reduce<Record<string, string>>((memo, attr) => {
+      value = element.getAttribute(attr);
+      if (value) {
+        memo[attr] = value;
+      }
+      return memo;
+    }, {}),
+    // add values parsed from style, which take precedence over attributes
+    // (see: http://www.w3.org/TR/SVG/styling.html#UsingPresentationAttributes)
+    ...getGlobalStylesForElement(element, cssRules),
+    ...parseStyleAttribute(element),
+  };
+
+  if (ownAttributes[cPath]) {
+    element.setAttribute(cPath, ownAttributes[cPath]);
   }
-  fontSize = parentFontSize =
-    parentAttributes.fontSize || DEFAULT_SVG_FONT_SIZE;
   if (ownAttributes[fSize]) {
     // looks like the minimum should be 9px when dealing with ems. this is what looks like in browsers.
-    ownAttributes[fSize] = fontSize = parseUnit(
-      ownAttributes[fSize],
-      parentFontSize
-    );
+    fontSize = parseUnit(ownAttributes[fSize], parentFontSize);
+    ownAttributes[fSize] = `${fontSize}`;
   }
 
-  const normalizedStyle = {};
+  const normalizedStyle: Record<string, string> = {};
   for (const attr in ownAttributes) {
     const normalizedAttr = normalizeAttr(attr);
     const normalizedValue = normalizeValue(