+++ src/shapes/Circle.ts
@@ -6,12 +6,13 @@ import { degreesToRadians } from '../util/misc/radiansDegreesConversion';
 import { sin } from '../util/misc/sin';
 import { classRegistry } from '../ClassRegistry';
 import { FabricObject, cacheProperties } from './Object/FabricObject';
-import type { TClassProperties } from '../typedefs';
+import type { Abortable, TClassProperties } from '../typedefs';
 import type {
   FabricObjectProps,
   SerializedObjectProps,
   TProps,
 } from './Object/types';
+import type { CSSRules } from '../parser/typedefs';
 
 interface UniqueCircleProps {
   /**
@@ -200,17 +201,25 @@ export class Circle<
    * Returns {@link Circle} instance from an SVG element
    * @static
    * @memberOf Circle
-   * @param {SVGElement} element Element to parse
+   * @param {HTMLElement} element Element to parse
    * @param {Object} [options] Partial Circle object to default missing properties on the element.
    * @throws {Error} If value of `r` attribute is missing or invalid
    */
-  static async fromElement(element: SVGElement): Promise<Circle> {
+  static async fromElement(
+    element: HTMLElement,
+    options: Abortable,
+    cssRules?: CSSRules
+  ): Promise<Circle> {
     const {
       left = 0,
       top = 0,
       radius = 0,
       ...otherParsedAttributes
-    } = parseAttributes(element, this.ATTRIBUTE_NAMES) as Partial<CircleProps>;
+    } = parseAttributes(
+      element,
+      this.ATTRIBUTE_NAMES,
+      cssRules
+    ) as Partial<CircleProps>;
 
     // this probably requires to be fixed for default origins not being top/left.