+++ src/shapes/Text/Text.ts
@@ -6,6 +6,7 @@ import { SHARED_ATTRIBUTES } from '../../parser/attributes';
 import { parseAttributes } from '../../parser/parseAttributes';
 import type { XY } from '../../Point';
 import type {
+  Abortable,
   TCacheCanvasDimensions,
   TClassProperties,
   TFiller,
@@ -39,6 +40,9 @@ import type {
 } from './TextStyles';
 import { TextStyles, hasStyleChanged } from './TextStyles';
 import { isFiller } from '../../util/typeAssertions';
+import type { Gradient } from '../../gradient/Gradient';
+import type { Pattern } from '../../Pattern';
+import type { CSSRules } from '../../parser/typedefs';
 
 let measuringContext: CanvasRenderingContext2D | null;
 
@@ -1854,11 +1858,19 @@ export class Text<
    * Returns Text instance from an SVG element (<b>not yet implemented</b>)
    * @static
    * @memberOf Text
-   * @param {SVGElement} element Element to parse
+   * @param {HTMLElement} element Element to parse
    * @param {Object} [options] Options object
    */
-  static async fromElement(element: SVGElement, options: object) {
-    const parsedAttributes = parseAttributes(element, Text.ATTRIBUTE_NAMES);
+  static async fromElement(
+    element: HTMLElement,
+    options: Abortable,
+    cssRules?: CSSRules
+  ) {
+    const parsedAttributes = parseAttributes(
+      element,
+      Text.ATTRIBUTE_NAMES,
+      cssRules
+    );
 
     const {
       textAnchor = LEFT as typeof LEFT | typeof CENTER | typeof RIGHT,